Consolidate five timing-safe compare copies into one shared module
Bearer-token and password checks were comparing secrets with five local
constant-time implementations. One shared primitive is one thing to audit.
