+++ packages/backup-control-plane/restore-confirm-token.ts
@@ -2,6 +2,7 @@ import {
 	base64UrlToBytes,
 	bytesToBase64Url,
 } from '@kody-internal/shared/base64.ts'
+import { timingSafeEqualString } from '@kody-internal/shared/timing-safe.ts'
 
 import { BackupError } from './backup-policy.ts'
 import { type BackupEnvironment } from './backup-types.ts'
@@ -33,17 +34,6 @@ async function hmacSha256(
 	return new Uint8Array(
 		await crypto.subtle.sign('HMAC', key, new TextEncoder().encode(message)),
 	)
-}
-
-export function constantTimeEqual(left: string, right: string): boolean {
-	const leftBytes = new TextEncoder().encode(left)
-	const rightBytes = new TextEncoder().encode(right)
-	if (leftBytes.byteLength !== rightBytes.byteLength) return false
-	let mismatch = 0
-	for (let index = 0; index < leftBytes.byteLength; index += 1) {
-		mismatch |= leftBytes[index]! ^ rightBytes[index]!
-	}
-	return mismatch === 0
 }
 
 export type RestoreConfirmToken = {
@@ -90,7 +80,7 @@ export async function verifyRestoreConfirmToken(
 	const expected = bytesToBase64Url(
 		await hmacSha256(secret, `restore:${input.day}:${input.expiresAt}`),
 	)
-	if (!constantTimeEqual(expected, input.token)) {
+	if (!(await timingSafeEqualString(expected, input.token))) {
 		throw new BackupError(
 			'restore-confirm-invalid',
 			'restore confirmation token is invalid',
