+++ packages/status/retrospective-maintenance.ts
@@ -3,6 +3,7 @@
  * Lives on the status worker so origin / APP_DB being down cannot block it.
  */
 
+import { timingSafeEqualString } from '@kody-internal/shared/timing-safe.ts'
 import {
 	parseIncidentRetrospectiveInput,
 	stampIncidentRetrospective,
@@ -24,39 +25,6 @@ function readBearerToken(request: Request) {
 	return auth?.startsWith('Bearer ')
 		? auth.slice('Bearer '.length).trim()
 		: null
-}
-
-function timingSafeEqualDigests(
-	left: ArrayBuffer,
-	right: ArrayBuffer,
-): boolean {
-	if (left.byteLength !== right.byteLength) return false
-	const subtleWithTiming = crypto.subtle as SubtleCrypto & {
-		timingSafeEqual?: (a: BufferSource, b: BufferSource) => boolean
-	}
-	if (typeof subtleWithTiming.timingSafeEqual === 'function') {
-		return subtleWithTiming.timingSafeEqual(left, right)
-	}
-	const a = new Uint8Array(left)
-	const b = new Uint8Array(right)
-	let diff = 0
-	for (let index = 0; index < a.length; index += 1) {
-		diff |= a[index]! ^ b[index]!
-	}
-	return diff === 0
-}
-
-/** SHA-256 both sides, then compare digests so length does not short-circuit. */
-export async function timingSafeEqualString(
-	left: string,
-	right: string,
-): Promise<boolean> {
-	const encoder = new TextEncoder()
-	const [leftDigest, rightDigest] = await Promise.all([
-		crypto.subtle.digest('SHA-256', encoder.encode(left)),
-		crypto.subtle.digest('SHA-256', encoder.encode(right)),
-	])
-	return timingSafeEqualDigests(leftDigest, rightDigest)
 }
 
 export function parseIncidentRetrospectivePath(
