+++ packages/worker/src/webhooks/crypto.ts
@@ -4,7 +4,7 @@ import {
 } from '@kody-internal/shared/base64.ts'
 import { toHex } from '@kody-internal/shared/hex.ts'
 import { sha256Hex } from '@kody-internal/shared/sha256.ts'
-import { timingSafeEqualString } from '#worker/maintenance-handler.ts'
+import { timingSafeEqualString } from '@kody-internal/shared/timing-safe.ts'
 import {
 	type WebhookHmacAlgorithm,
 	type WebhookSignatureEncoding,
